Clarify the content model
This change is largely, but not entirely, editorial.

The "content model" (second normative paragraph in 2.1 and the
subsequent list) and the definition of ruby segments in 2.1.1 were
partially (but not entirely) redundant with each other. This changes
merges both, and uses that as a definition of the content model.

Along the way, this resolves a contradiction as to whether "one or more"
or "zero or more" of various pieces of the ruby model are allowed.

Finally, this change also makes segmentation a little more explicit, not
only specifying that each base range is associated with some annotation
ranges, but being explicit about which.

Finally, example #ruby-inlining (previous ex. 5, now ex. 3) is
unchanged, but moved around to attach to the most relevant part of the
reorganized text.